refactor: extract useDropdown hook from HelpDialog and UserMenu
Shared open/close state, click-outside dismiss, and Escape key handling.
