+++ packages/client/src/components/UserMenu.tsx
@@ -1,6 +1,6 @@
-import { useState, useRef, useEffect, useCallback } from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCaretDown } from '@fortawesome/free-solid-svg-icons';
+import useDropdown from '../hooks/useDropdown';
 import './UserMenu.css';
 
 interface UserMenuProps {
@@ -8,33 +8,13 @@ interface UserMenuProps {
 }
 
 export default function UserMenu({ onLogout }: UserMenuProps) {
-  const [open, setOpen] = useState<boolean>(false);
-  const menuRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    const handleClickOutside = (e: MouseEvent): void => {
-      if (menuRef.current && e.target instanceof Node && !menuRef.current.contains(e.target)) {
-        setOpen(false);
-      }
-    };
-    document.addEventListener('mousedown', handleClickOutside);
-    return () => document.removeEventListener('mousedown', handleClickOutside);
-  }, []);
-
-  const handleKeyDown = useCallback((e: React.KeyboardEvent) => {
-    if (e.key === 'Escape') setOpen(false);
-    if (e.key === 'Enter' || e.key === ' ') {
-      e.preventDefault();
-      setOpen((o) => !o);
-    }
-  }, []);
+  const { open, ref, toggle, handleKeyDown } = useDropdown();
 
   return (
-    <div className="user-menu" ref={menuRef}>
+    <div className="user-menu" ref={ref} onKeyDown={handleKeyDown}>
       <button
         className="user-menu__trigger btn-unstyled icon-interactive focus-ring"
-        onClick={() => setOpen(!open)}
-        onKeyDown={handleKeyDown}
+        onClick={toggle}
         aria-label="User menu"
         aria-expanded={open}
         aria-haspopup="true"
